Fixed wrong typename in UIntColumnType.

+++ src/Pure.RelationalSchema/ColumnType/UIntColumnType.cs
@@ -6,7 +6,7 @@ namespace Pure.RelationalSchema.ColumnType;
 
 public sealed record UIntColumnType : IColumnType
 {
-    IString IColumnType.Name { get; } = new String("ulong");
+    IString IColumnType.Name { get; } = new String("uint");
 
     public override int GetHashCode()
     {
